requirements.txt
# Google Calendar API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# AI Integration
google-generativeai==0.3.2

# Telegram Bot
python-telegram-bot==20.7

# Configuration and Environment
python-dotenv==1.0.0

# Scheduling
schedule==1.2.0

# Database
# SQLite3 is included in Python standard library

# Utilities
requests==2.31.0
pytz==2023.3
